requirements.txt
Django==4.1.4
djangorestframework==3.14.0 
Pillow==9.3.0
python-slugify==7.0.0
django-cors-headers==3.13.0
django-filter==22.1